+++ app/src/main/java/com/zhuravlevmikhail/a65appshomeproject/core/contacts/ContactsContract.kt
@@ -1,6 +1,5 @@
-package com.zhuravlevmikhail.a65appshomeproject.core.fragm_contact
+package com.zhuravlevmikhail.a65appshomeproject.core.contacts
 
-import android.content.AsyncQueryHandler
 import android.content.ContentResolver
 import com.zhuravlevmikhail.a65appshomeproject.core.mvpAchitecture.MvpModel
 import com.zhuravlevmikhail.a65appshomeproject.core.mvpAchitecture.MvpPresenter
@@ -9,17 +8,17 @@ import io.reactivex.Observable
 
 interface ContactsContract {
 
-    interface TModel : MvpModel
+    interface ContactsModelContract : MvpModel
 
-    interface TPresenter<V : TView> : MvpPresenter<V> {
-        fun queryContactsAsync(
-            contentResolver: ContentResolver
-        ): Observable<ArrayList<ContactsModel.ContactGeneral>>
-    }
-
-    interface TView : MvpView {
+    interface ContactsViewContract : MvpView {
         fun onContactsAccessGranted()
         fun setContacts(newContacts: ArrayList<ContactsModel.ContactGeneral>)
         fun openDetailedContactPage(contactId: Long)
+    }
+
+    interface ContactsPresenterContract<V : ContactsViewContract> : MvpPresenter<V> {
+        fun queryContactsAsync(
+            contentResolver: ContentResolver
+        ): Observable<ArrayList<ContactsModel.ContactGeneral>>
     }
 }
