+++ app/src/main/java/com/zhuravlevmikhail/a65appshomeproject/core/detail/DetailedContract.kt
@@ -1,7 +1,6 @@
-package com.zhuravlevmikhail.a65appshomeproject.core.fragm_con_detailed
+package com.zhuravlevmikhail.a65appshomeproject.core.detail
 
 import android.content.ContentResolver
-import android.graphics.Bitmap
 import com.zhuravlevmikhail.a65appshomeproject.core.mvpAchitecture.MvpModel
 import com.zhuravlevmikhail.a65appshomeproject.core.mvpAchitecture.MvpPresenter
 import com.zhuravlevmikhail.a65appshomeproject.core.mvpAchitecture.MvpView
@@ -9,15 +8,11 @@ import io.reactivex.Observable
 
 interface DetailedContract {
 
-    interface TModel : MvpModel
+    interface DetailedModelContract : MvpModel
 
-    interface TView : MvpView
+    interface DetailedViewContract : MvpView
 
-    interface TPresenter<V : TView> : MvpPresenter<V> {
+    interface DetailedPresenterContract<V : DetailedViewContract> : MvpPresenter<V> {
         fun queryContactWithoutImageAsync(contentResolver: ContentResolver, contactId: Long): Observable<DetailedModel.ContactDetailed>
-        /*fun queryContactImageAsync(
-            contentResolver: ContentResolver,
-            contactId: Long
-        ): Observable<Bitmap>*/
     }
 }
